[Chore #161203714] Add tests for get all products
